requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
opencv-python>=4.5.5
pillow>=9.0.0
timm>=0.9.2
tqdm>=4.60.0
PyQt5>=5.15.0
psutil>=5.8.0
